extract all components from main.rs
